Move code dealing with QR Code payload to separate package
