+++ feature/migration/qrcode/src/test/kotlin/app/k9mail/feature/migration/qrcode/payload/QrCodePayloadReaderTest.kt
@@ -1,4 +1,4 @@
-package app.k9mail.feature.migration.qrcode
+package app.k9mail.feature.migration.qrcode.payload
 
 import app.k9mail.core.common.mail.toUserEmailAddress
 import app.k9mail.core.common.net.toHostname
